requirements.txt
playwright
pytest
pytest-playwright
python-dotenv
gspread
google-auth
Flask
waitress
pyngrok
openai
